pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lingq"
version = "1.0.0"
readme = "README.md"
dependencies = [
    "aiohttp==3.9.5",
    "aiohttp_retry==2.8.3",
    "click==8.1.7",
    "natsort==8.4.0",
    "python-dotenv==1.0.1",
    "yt_dlp==2024.10.22",
    "pydantic",
    "roman==4.2",
]

[project.optional-dependencies]
etc = [
    "beautifulsoup4==4.12.3",
    "faster_whisper==1.0.2",
    "requests==2.32.3",
    "ruff==0.7.0",
    "tqdm==4.66.4",
    "gpt4All",
]
test = [
    "pytest==8.2.1",
    "deepdiff",
]
all = ["lingq[etc, test]"]

[project.scripts]
lingq = "cli:cli"

[tool.ruff]
line-length = 100
output-format = "concise"
exclude = [
    "etc/audio_processing/split_audio.py",
]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint]
preview = true
select = [
    "W",    # pycodestyle warning
    "I",    # isort
    "N",    # pep8-naming
    "F",    # Pyflakes rules
    "C",    # flake8-complexity
    "PTH",  # Paths
    "ANN",  # Annotations
]
ignore = [
    "ANN401",
]

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
